- added 'general' state - moved shards to 'general' state
