+++ src/hooks/fetch/useFetchShards.ts
@@ -2,15 +2,15 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { useAdapter } from 'hooks';
-import { interfaceSelector } from 'redux/selectors';
-import { setShards } from 'redux/slices/interface';
+import { generalSelector } from 'redux/selectors';
+import { setShards } from 'redux/slices/general';
 
 let currentRequest: any = null;
 
 export const useFetchShards = () => {
   const dispatch = useDispatch();
   const { getShards } = useAdapter();
-  const { shards } = useSelector(interfaceSelector);
+  const { shards } = useSelector(generalSelector);
 
   const getShardsOnce = () => {
     if (currentRequest) {
