+++ src/redux/selectors/index.ts
@@ -3,7 +3,7 @@ export * from './accountExtra';
 export * from './accountStaking';
 export * from './collection';
 export * from './economics';
-export * from './stake';
+export * from './general';
 export * from './growthHero';
 export * from './growthMostUsed';
 export * from './growthPrice';
@@ -19,5 +19,6 @@ export * from './nft';
 export * from './nodesIdentities';
 export * from './nodesVersions';
 export * from './refresh';
+export * from './stake';
 export * from './stats';
 export * from './token';
