+++ src/redux/slices/interface.ts
@@ -1,24 +1,14 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import moment from 'moment';
 
-import {
-  METACHAIN_SHARD_ID,
-  TEMP_LOCAL_NOTIFICATION_DISMISSED
-} from 'appConstants';
+import { TEMP_LOCAL_NOTIFICATION_DISMISSED } from 'appConstants';
 import { storage } from 'helpers';
-import { sortShards } from 'helpers/sortShards';
-import {
-  ExplorerOriginType,
-  ThemesEnum,
-  NotificationType,
-  ShardType
-} from 'types';
+import { ExplorerOriginType, ThemesEnum, NotificationType } from 'types';
 
 export type InterfaceSliceType = {
   activeTheme: ThemesEnum;
   explorerOrigin: ExplorerOriginType;
   notifications: NotificationType[];
-  shards: ShardType[];
 };
 
 export function getInitialInterfaceState(): InterfaceSliceType {
@@ -36,8 +26,7 @@ export function getInitialInterfaceState(): InterfaceSliceType {
       pathname: '/',
       search: ''
     },
-    notifications: [],
-    shards: []
+    notifications: []
   };
 }
 
@@ -56,12 +45,6 @@ export const interfaceSlice = createSlice({
       action: PayloadAction<InterfaceSliceType['explorerOrigin']>
     ) => {
       state.explorerOrigin = action.payload;
-    },
-    setShards: (
-      state: InterfaceSliceType,
-      action: PayloadAction<InterfaceSliceType['shards']>
-    ) => {
-      state.shards = sortShards({ shards: action.payload, METACHAIN_SHARD_ID });
     },
     addNotification: (
       state: InterfaceSliceType,
@@ -98,7 +81,6 @@ export const interfaceSlice = createSlice({
 export const {
   setActiveTheme,
   setExplorerOrigin,
-  setShards,
   addNotification,
   removeNotification
 } = interfaceSlice.actions;
